app-launch-context: Rename files to follow the prefix-name rule
Moves the files from launch-context to xdp-app-launch-context because it
implements the XdpAppLaunchContext class.
